fix(jangar): refresh architecture inventory
